backend/pyproject.toml
[project]
name = "fastapi-neo4j"
version = "0.1.0"
description = "A simple starter for building RESTful APIs with FastAPI and NEO4J."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.115.13",
    "httpx>=0.28.1",
    "neontology>=2.1.0",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.7",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.8.0",
    "ty>=0.0.21",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
